Add files via upload
M&A analysis using Accretion/Dilution Analysis to assesses whether the acquiring company’s earnings per share (EPS) will increase or decrease posttransaction. This model is crucial for evaluating the financial impact of a merger or acquisition, helping stakeholders understand the transaction's value and feasibility. By applying this analysis to Google's potential acquisition of Motorola, a positive accretion of $0.90 per share was identified, indicating a 3.6 percent increase in EPS. This positive result not only demonstrates the financial viability of the deal but also signals that the transaction is likely to enhance shareholders value. Therefore, the model provides valuable insights for making informed decisions and assessing the success of M&A activities.